Movie information and Cinema venue.
This App will help a Customer to view Whats Up at the cinema.
Customer will be able to buy/book tickets online with a few clicks.
The customer will get a google map to get to the cinema. Which bus/train to catch to get there.
Duration of the movie and what time it will end.